builder: Avoid rewriting -savemanifest during release phase
If the build runs during the release phase, -savemanifest would store
the manifest from the rebuilt jar with a potentially later time stamp.
We only save the manifest if the built jar timestamp is newer than the
output manifest file.

+++ biz.aQute.bndlib/src/aQute/bnd/osgi/Builder.java
@@ -303,18 +303,17 @@ public class Builder extends Analyzer {
 		if (f.isDirectory()) {
 			f = new File(f, "MANIFEST.MF");
 		}
-		f.delete();
-		File fp = f.getParentFile();
-		if (!fp.exists() && !fp.mkdirs()) {
-			throw new IOException("Could not create directory " + fp);
+		if (!f.exists() || f.lastModified() < dot.lastModified()) {
+			f.delete();
+			File fp = f.getParentFile();
+			if (!fp.exists() && !fp.mkdirs()) {
+				throw new IOException("Could not create directory " + fp);
+			}
+			try (OutputStream out = new FileOutputStream(f)) {
+				Jar.writeManifest(dot.getManifest(), out);
+			}
+			changedFile(f);
 		}
-		OutputStream out = new FileOutputStream(f);
-		try {
-			Jar.writeManifest(dot.getManifest(), out);
-		} finally {
-			out.close();
-		}
-		changedFile(f);
 	}
 
 	protected void changedFile(@SuppressWarnings("unused") File f) {}
